combining the file processing functions
